rad-clone: Initialize
Big refactor as well.
