+++ auth/src/lib.rs
@@ -1,14 +1,18 @@
-use rad_terminal::components::{Args, Error, Help};
+use std::ffi::OsString;
+
+use rad_terminal::args::{Args, Error, Help};
 
 pub const HELP: Help = Help {
     name: "auth",
     description: env!("CARGO_PKG_DESCRIPTION"),
     version: env!("CARGO_PKG_VERSION"),
     usage: r#"
-USAGE
+Usage
+
     rad auth [--init]
 
-OPTIONS
+Options
+
     --init    Initialize a new identity
     --help    Print help
 "#,
@@ -20,11 +24,11 @@ pub struct Options {
 }
 
 impl Args for Options {
-    fn from_env() -> Result<Self, anyhow::Error> {
+    fn from_args(args: Vec<OsString>) -> anyhow::Result<(Self, Vec<OsString>)> {
         use lexopt::prelude::*;
 
         let mut init = false;
-        let mut parser = lexopt::Parser::from_env();
+        let mut parser = lexopt::Parser::from_args(args);
 
         while let Some(arg) = parser.next()? {
             match arg {
@@ -38,6 +42,6 @@ impl Args for Options {
             }
         }
 
-        Ok(Options { init })
+        Ok((Options { init }, vec![]))
     }
 }
