+++ clone/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "rad-clone"
+version = "0.1.0"
+authors = ["The Radicle Team <dev@radicle.xyz>"]
+edition = "2018"
+license = "MIT OR Apache-2.0"
+description = "Clone radicle projects"
+
+[dependencies]
+anyhow = "1.0"
+librad = "0"
+lexopt = "0.2"
+rad-terminal = { path = "../terminal" }
+rad-common = { path = "../common" }
+rad-sync = { path = "../sync" }
+rad-checkout = { path = "../checkout" }
+rad-track = { path = "../track" }
