+++ common/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod args;
 pub mod ethereum;
 pub mod git;
 pub mod keys;
