+++ help/Cargo.toml
@@ -20,6 +20,7 @@ rad-track = { path = "../track" }
 rad-untrack = { path = "../untrack" }
 rad-init = { path = "../init" }
 rad-auth = { path = "../auth" }
+rad-clone = { path = "../clone" }
 rad-account = { path = "../account" }
 rad-checkout = { path = "../checkout" }
 rad-push = { path = "../push" }
