+++ src/bin/rad-account.rs
@@ -1,8 +1,8 @@
 use rad_account::{run, Options, HELP};
-use rad_terminal::components as term;
+use rad_terminal::args;
 
 fn main() {
-    term::run_command::<Options, _>(HELP, "Command", move |opts| {
+    args::run_command::<Options, _>(HELP, "Command", move |opts| {
         let rt = tokio::runtime::Runtime::new()?;
         rt.block_on(run(opts))
     });
