+++ src/bin/rad-clone.rs
@@ -1,0 +1,6 @@
+use rad_clone::{run, Options, HELP};
+use rad_terminal::args;
+
+fn main() {
+    args::run_command::<Options, _>(HELP, "Cloning", run);
+}
