+++ src/bin/rad-ens.rs
@@ -1,6 +1,6 @@
 use rad_ens::{run, Options, HELP};
-use rad_terminal::components as term;
+use rad_terminal::args;
 
 fn main() {
-    term::run_command::<Options, _>(HELP, "Command", run);
+    args::run_command::<Options, _>(HELP, "Command", run);
 }
