+++ src/bin/rad-ls.rs
@@ -1,6 +1,6 @@
 use rad_ls::{run, Options, HELP};
-use rad_terminal::components as term;
+use rad_terminal::args;
 
 fn main() {
-    term::run_command::<Options, _>(HELP, "Listing objects", run);
+    args::run_command::<Options, _>(HELP, "Listing objects", run);
 }
