+++ src/bin/rad-sync.rs
@@ -1,6 +1,6 @@
 use rad_sync::{run, Options, HELP};
-use rad_terminal::components as term;
+use rad_terminal::args;
 
 fn main() {
-    term::run_command::<Options, _>(HELP, "Sync", run);
+    args::run_command::<Options, _>(HELP, "Sync", run);
 }
