+++ src/bin/rad-untrack.rs
@@ -1,7 +1,7 @@
-use rad_terminal::components as term;
-use rad_track::options::Options;
+use rad_terminal::args;
+use rad_track::Options;
 use rad_untrack::{run, HELP};
 
 fn main() {
-    term::run_command::<Options, _>(HELP, "Untracking", run);
+    args::run_command::<Options, _>(HELP, "Untracking", run);
 }
